Refactoring: Move and split Workerify
